Added Goblin Blastmage combo
Goblin Blastmage is less likely to be cast, in the case of no mechs being on the field, and there being at least 1 mech in your hand. If the mech is playable at the same turn, the mech will be cast first, and then Goblin Blastmage. Blastmage will be played, if there are absolutely no worthy minions to play.

I don't have Goblin Blastmage to test this code.